+++ src/__tests__/didSupport.test.js
@@ -1,6 +1,10 @@
 const OrbitDB = require('orbit-db')
+const { LegacyIPFS3BoxAccessController } = require('3box-orbitdb-plugins')
+const AccessControllers = require('orbit-db-access-controllers')
+AccessControllers.addAccessController({ AccessController: LegacyIPFS3BoxAccessController })
 const Util = require('../util')
 const { makeIPFS } = require('./tools')
+const registerMuportResolver = require('muport-did-resolver')
 
 const IPFS_PATH = './tmp/ipfs-did-1'
 const ODB_PATH = './tmp/orbitdb-did-1'
@@ -34,8 +38,9 @@ describe('basic low level functions are working', () => {
   })
 
   test('did extract signing key', async () => {
-    const ipfs = { files: { cat: jest.fn(() => Promise.resolve(MANIFEST)) } }
-    const k = await Util.didExtractSigningKey(ipfs, DID)
+    const ipfs = { cat: jest.fn(() => Promise.resolve(MANIFEST)) }
+    registerMuportResolver(ipfs)
+    const k = await Util.didExtractSigningKey(DID)
     expect(k).toEqual('02d1f48e3d5c52954a01f1aa104bad1a22e2eed6ecbd4961737fbffa8d75457cd4')
   })
 })
@@ -47,7 +52,7 @@ describe('test with network', () => {
 
   beforeAll(async () => {
     ipfs = await makeIPFS(IPFS_CONF)
-    const res = await ipfs.files.add(Buffer.from(MANIFEST))
+    const res = await ipfs.add(Buffer.from(MANIFEST))
     orbitdb = new OrbitDB(ipfs, ODB_PATH)
   })
 
@@ -56,7 +61,7 @@ describe('test with network', () => {
   })
 
   it('can retrieve a root store', async () => {
-    const addr = await Util.didToRootStoreAddress(DID, { ipfs, orbitdb })
+    const addr = await Util.didToRootStoreAddress(DID, { orbitdb })
     expect(addr).toEqual(ROOT_STORE_ADDR)
   })
 })